+++ Kernel/Modules/AgentTicketQueue.pm
@@ -607,6 +607,7 @@ sub _MaskQueueView {
     }
 
     # build queue string
+    QUEUE:
     for my $QueueRef (@ListedQueues) {
         my $QueueStrg = '';
         my %Queue     = %$QueueRef;
@@ -621,6 +622,17 @@ sub _MaskQueueView {
         my $ShortQueueName = $QueueName[-1];
         $Queue{MaxAge} = $Queue{MaxAge} / 60;
         $Queue{QueueID} = 0 if ( !$Queue{QueueID} );
+
+        # skip empty Queues (or only locked tickets)
+        if (
+            $Counter{ $Queue{Queue} } < 1
+            && $Queue{Queue} ne $LayoutObject->{LanguageObject}->Translate('My Queues')
+            && $Config->{HideEmptyQueues}
+            )
+        {
+            # TODO: check what 'Ticket::ViewableLocks' affects
+            next QUEUE;
+        }
 
         my $View   = $Kernel::OM->Get('Kernel::System::Web::Request')->GetParam( Param => 'View' )   || '';
         my $Filter = $Kernel::OM->Get('Kernel::System::Web::Request')->GetParam( Param => 'Filter' ) || 'Unlocked';